refactor: split long types.ts file into multiple files on lib/types/
