+++ src/lib/types/api.ts
@@ -1,0 +1,7 @@
+export type User = {
+	id: string;
+	firstName: string;
+	lastName: string;
+	email: string;
+	creationDate: Date;
+};
