+++ src/lib/types/schemas.ts
@@ -1,30 +1,5 @@
 import z from "zod";
 
-export enum SessionStatus {
-	AUTHENTICATED = "authenticated",
-	NO_SESSION = "no_session",
-}
-
-export type ErrorEntry = {
-	code: string;
-	field?: string; // for the case of using it as an field error
-	message: string;
-	statusCode: number;
-};
-
-export type FieldError = {
-	type: string;
-	value: string;
-	msg: string;
-	path: string;
-	location: string;
-};
-
-export type FormattedFieldError = {
-	path: string;
-	messages: string[];
-};
-
 export const REGISTER_FORM_SCHEMA = z
 	.object({
 		firstName: z
