+++ src/lib/types/types.ts
@@ -1,0 +1,4 @@
+export enum SessionStatus {
+	AUTHENTICATED = "authenticated",
+	NO_SESSION = "no_session",
+}
